Refactor HotTake component to SpicyPage
